#54 feat: move all resource files by the category path name
