+++ src/Http/Controllers/CategoryController.php
@@ -7,7 +7,7 @@ use Botble\Blog\Http\Resources\CategoryResource;
 use Botble\Blog\Repositories\Interfaces\CategoryInterface;
 use Botble\Blog\Supports\FilterCategory;
 use CSlant\Blog\Api\Enums\StatusEnum;
-use CSlant\Blog\Api\Http\Resources\ListCategoryResource;
+use CSlant\Blog\Api\Http\Resources\Category\ListCategoryResource;
 use CSlant\Blog\Api\OpenApi\Schemas\Resources\Category\CategoryListResourceSchema;
 use CSlant\Blog\Api\OpenApi\Schemas\Resources\Category\CategoryModelResourceSchema;
 use CSlant\Blog\Core\Facades\Base\SlugHelper;
