+++ src/Http/Controllers/TagController.php
@@ -4,7 +4,7 @@ namespace CSlant\Blog\Api\Http\Controllers;
 
 use Botble\Base\Http\Responses\BaseHttpResponse;
 use CSlant\Blog\Api\Enums\StatusEnum;
-use CSlant\Blog\Api\Http\Resources\TagResource;
+use CSlant\Blog\Api\Http\Resources\Tag\TagResource;
 use CSlant\Blog\Api\OpenApi\Responses\Errors\BadRequestResponseSchema;
 use CSlant\Blog\Api\OpenApi\Responses\Errors\ErrorNotFoundResponseSchema;
 use CSlant\Blog\Api\OpenApi\Responses\Errors\InternalServerResponseSchema;
