+++ src/Providers/BlogApiServiceProvider.php
@@ -2,7 +2,6 @@
 
 namespace CSlant\Blog\Api\Providers;
 
-use Illuminate\Http\Resources\Json\JsonResource;
 use Illuminate\Support\ServiceProvider;
 
 class BlogApiServiceProvider extends ServiceProvider
@@ -81,7 +80,7 @@ class BlogApiServiceProvider extends ServiceProvider
     {
         if (!class_exists(\Botble\Blog\Http\Resources\TagResource::class, false)) {
             class_alias(
-                \CSlant\Blog\Api\Http\Resources\TagResource::class,
+                \CSlant\Blog\Api\Http\Resources\Tag\TagResource::class,
                 \Botble\Blog\Http\Resources\TagResource::class
             );
         }
